[DRAFT] Message decoder (#780)
Implemented decoder and moved helper functions to serialization_strategy.
